Split VM / Nav API - Retrograde to composeNavigation = "2.8.0-alpha02"
