+++ projects/compose/koin-compose-viewmodel-navigation/build.gradle.kts
@@ -36,9 +36,9 @@ kotlin {
 
     sourceSets {
         commonMain.dependencies {
-            api(project(":compose:koin-compose"))
-            api(project(":core:koin-core-viewmodel"))
-            api(libs.jb.composeViewmodel)
+            api(project(":compose:koin-compose-viewmodel"))
+            api(project(":core:koin-core-viewmodel-navigation"))
+            api(libs.jb.composeNavigation)
         }
     }
 }
