+++ projects/core/koin-core-viewmodel-navigation/build.gradle.kts
@@ -4,8 +4,6 @@ import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 
 plugins {
     alias(libs.plugins.kotlinMultiplatform)
-    alias(libs.plugins.jetbrainsCompose)
-    alias(libs.plugins.compose.compiler)
 }
 
 val koinVersion: String by project
@@ -24,8 +22,8 @@ kotlin {
     }
 
     wasmJs {
-        nodejs()
         binaries.executable()
+        nodejs()
     }
 
     iosX64()
@@ -36,9 +34,8 @@ kotlin {
 
     sourceSets {
         commonMain.dependencies {
-            api(project(":compose:koin-compose"))
             api(project(":core:koin-core-viewmodel"))
-            api(libs.jb.composeViewmodel)
+            api(libs.jb.navigation)
         }
     }
 }
@@ -49,8 +46,10 @@ tasks.withType<KotlinCompile>().all {
     }
 }
 
-tasks.withType<org.jetbrains.kotlin.gradle.targets.js.npm.tasks.KotlinNpmInstallTask>().configureEach {
-    args.add("--ignore-engines")
+tasks.withType<KotlinCompile>().all {
+    compilerOptions {
+        jvmTarget.set(JvmTarget.JVM_1_8)
+    }
 }
 
 apply(from = file("../../gradle/publish.gradle.kts"))
