+++ projects/core/koin-core-viewmodel-navigation/src/commonMain/kotlin/org/koin/viewmodel/CreationExtrasExt.kt
@@ -1,0 +1,14 @@
+package org.koin.viewmodel
+
+import androidx.lifecycle.HasDefaultViewModelProviderFactory
+import androidx.lifecycle.ViewModelStoreOwner
+import androidx.lifecycle.viewmodel.CreationExtras
+import androidx.navigation.NavBackStackEntry
+import org.koin.core.annotation.KoinInternalApi
+
+@OptIn(KoinInternalApi::class)
+fun defaultNavExtras(viewModelStoreOwner: ViewModelStoreOwner): CreationExtras = when {
+    viewModelStoreOwner is NavBackStackEntry && viewModelStoreOwner.arguments != null -> viewModelStoreOwner.arguments?.toExtras(viewModelStoreOwner) ?: CreationExtras.Empty
+    viewModelStoreOwner is HasDefaultViewModelProviderFactory -> viewModelStoreOwner.defaultViewModelCreationExtras
+    else -> CreationExtras.Empty
+}
