+++ projects/gradle/libs.versions.toml
@@ -13,17 +13,17 @@ uuid = "0.8.4"
 # Android
 agp = "7.4.2"
 android-appcompat = "1.7.0"
-android-activity = "1.9.0"
-android-fragment = "1.8.1"
-androidx-lifecycle = "2.8.3"
-androidx-workmanager = "2.9.0"
+android-activity = "1.9.1"
+android-fragment = "1.8.2"
+androidx-lifecycle = "2.8.4"
+androidx-workmanager = "2.9.1"
 androidx-navigation = "2.7.7"
 
 # Compose
+composeJetpackRuntime = "1.6.8"
 composeJB = "1.6.11"
 composeLifecycle = "2.8.0"
-composeNavigation = "2.8.0-alpha08"
-composeJetpackRuntime = "1.6.8"
+composeNavigation = "2.8.0-alpha02" # 2.8.0-alpha02 to keep compose 1.6.11
 jbCoreBundle = "1.0.0"
 jbSavedState = "1.2.0"
 
@@ -64,18 +64,19 @@ ktor-core = { module = "io.ktor:ktor-server-core", version.ref = "ktor" }
 ktor-netty = { module = "io.ktor:ktor-server-netty", version.ref = "ktor" }
 ktor-testHost = { module = "io.ktor:ktor-server-test-host", version.ref = "ktor" }
 ktor-slf4j = { module = "org.slf4j:slf4j-api", version.ref = "slf4j" }
-# Compose
+# jetpack Compose
+androidx-composeRuntime = { module = "androidx.compose.runtime:runtime", version.ref = "composeJetpackRuntime" }
+androidx-composeViewModel = { module = "androidx.lifecycle:lifecycle-viewmodel-compose", version.ref = "androidx-lifecycle" }
+androidx-composeNavigation = { module = "androidx.navigation:navigation-compose", version.ref = "androidx-navigation" }
+# jb Compose
 jb-composeRuntime = { module = "org.jetbrains.compose.runtime:runtime", version.ref = "composeJB" }
 jb-composeViewmodel = { module = "org.jetbrains.androidx.lifecycle:lifecycle-viewmodel-compose", version.ref = "composeLifecycle" }
-jb-composeNavigation = { module = "org.jetbrains.androidx.navigation:navigation-compose", version.ref = "composeNavigation" }
 jb-lifecycleViewmodel = { module = "org.jetbrains.androidx.lifecycle:lifecycle-viewmodel", version.ref = "composeLifecycle" }
 jb-lifecycleViewmodelSavedState = { module = "org.jetbrains.androidx.lifecycle:lifecycle-viewmodel-savedstate", version.ref = "composeLifecycle" }
 jb-bundle = { module = "org.jetbrains.androidx.core:core-bundle", version.ref = "jbCoreBundle" }
 jb-savedstate = { module = "org.jetbrains.androidx.savedstate:savedstate", version.ref = "jbSavedState" }
+jb-composeNavigation = { module = "org.jetbrains.androidx.navigation:navigation-compose", version.ref = "composeNavigation" }
 jb-navigation = { module = "org.jetbrains.androidx.navigation:navigation-common", version.ref = "composeNavigation" }
-androidx-composeRuntime = { module = "androidx.compose.runtime:runtime", version.ref = "composeJetpackRuntime" }
-androidx-composeViewModel = { module = "androidx.lifecycle:lifecycle-viewmodel-compose", version.ref = "androidx-lifecycle" }
-androidx-composeNavigation = { module = "androidx.navigation:navigation-compose", version.ref = "androidx-navigation" }
 
 [plugins]
 kotlinMultiplatform = { id = "org.jetbrains.kotlin.multiplatform", version.ref = "kotlin" }
